backend/requirements.txt
fastapi[standard]
python-dotenv
httpx
aiofiles
requests
